Add initial configuration and main server setup
This commit includes the initial setup and configuration of the TypeScript project with linting and TailwindCSS. A `tsconfig.json` has been added along with necessary `.eslintrc.json`, `postcss.config.js`, and `tailwind.config.ts` files. These contain necessary rules and configurations aligning with Next.js  and react linting standards and Tailwind CSS styling.

`.gitignore` files for both global and webapp specific contexts have also been added to ignore unnecessary files/directories.

A basic GRPC server setup is added in `server/main.go` and the necessary routes are defined in `protos/example.proto` for communication.

A `package.json` file is also added along with `package-lock.json` in the webapp directory indicating the necessary dependencies for the frontend application.

This commit is necessary to lay the foundation needed for development of the frontend and backend code.